Feat: ImageCache을 uploader로 분리
